+++ src/components/ui/form/DialogueTextarea.tsx
@@ -69,7 +69,7 @@ const DialogueTextareaWrapper = styled.div`
   .MentionsInput__highlighter {
     color: ${(props) => props.theme.colors.input.text} !important;
     font-family: monospace;
-    font-size: 12px;
+    font-size: ${(props) => props.theme.typography.fontSize};
     font-stretch: 100%;
     font-style: normal;
     font-variant-caps: normal;
@@ -132,6 +132,7 @@ const DialogueTextareaWrapper = styled.div`
     display: flex;
     align-items: center;
     padding: 5px 10px;
+    font-size: ${(props) => props.theme.typography.menuFontSize};
     &:focus {
       background: ${(props) => props.theme.colors.menu.hoverBackground};
       outline: none;