+++ src/lib/events/eventIfActorAtPosition.js
@@ -3,11 +3,13 @@ const l10n = require("../helpers/l10n").default;
 const id = "EVENT_IF_ACTOR_AT_POSITION";
 const groups = ["EVENT_GROUP_CONTROL_FLOW", "EVENT_GROUP_ACTOR"];
 
-const autoLabel = (fetchArg) => {
+const autoLabel = (fetchArg, input) => {
+  const unitPostfix =
+    input.units === "pixels" ? l10n("FIELD_PIXELS_SHORT") : "";
   return l10n("EVENT_IF_ACTOR_AT_POSITION_LABEL", {
     actor: fetchArg("actorId"),
-    x: fetchArg("x"),
-    y: fetchArg("y"),
+    x: `${fetchArg("x")}${unitPostfix}`,
+    y: `${fetchArg("y")}${unitPostfix}`,
   });
 };
 
@@ -19,22 +21,33 @@ const fields = [
     defaultValue: "$self$",
   },
   {
-    key: "x",
-    label: l10n("FIELD_X"),
-    type: "number",
-    min: 0,
-    max: 255,
-    width: "50%",
-    defaultValue: 0,
-  },
-  {
-    key: "y",
-    label: l10n("FIELD_Y"),
-    type: "number",
-    min: 0,
-    max: 255,
-    width: "50%",
-    defaultValue: 0,
+    type: "group",
+    fields: [
+      {
+        key: "x",
+        label: l10n("FIELD_X"),
+        type: "number",
+        min: 0,
+        max: 255,
+        width: "50%",
+        defaultValue: 0,
+        unitsField: "units",
+        unitsDefault: "tiles",
+        unitsAllowed: ["tiles", "pixels"],
+      },
+      {
+        key: "y",
+        label: l10n("FIELD_Y"),
+        type: "number",
+        min: 0,
+        max: 255,
+        width: "50%",
+        defaultValue: 0,
+        unitsField: "units",
+        unitsDefault: "tiles",
+        unitsAllowed: ["tiles", "pixels"],
+      },
+    ],
   },
   {
     key: "true",
@@ -75,7 +88,7 @@ const compile = (input, helpers) => {
   const truePath = input.true;
   const falsePath = input.__disableElse ? [] : input.false;
   actorSetActive(input.actorId);
-  ifActorAtPosition(input.x, input.y, truePath, falsePath);
+  ifActorAtPosition(input.x, input.y, truePath, falsePath, input.units);
 };
 
 module.exports = {