+++ src/store/features/entities/entitiesTypes.ts
@@ -457,6 +457,17 @@ export interface ScriptEventFieldCondition {
   soundType?: unknown;
 }
 
+export const distanceUnitTypes = ["tiles", "pixels"] as const;
+export const timeUnitTypes = ["time", "frames"] as const;
+export const unitTypes = [...distanceUnitTypes, ...timeUnitTypes] as const;
+
+export type UnitType = typeof unitTypes[number];
+export type DistanceUnitType = typeof distanceUnitTypes[number];
+export type TimeUnitType = typeof timeUnitTypes[number];
+
+export const movementTypes = ["horizontal", "vertical", "diagonal"] as const;
+export type MovementType = typeof movementTypes[number];
+
 export interface ScriptEventFieldSchema {
   label?: string | React.ReactNode;
   checkboxLabel?: string;
@@ -468,7 +479,8 @@ export interface ScriptEventFieldSchema {
   conditions?: ScriptEventFieldCondition[];
   toggleLabel?: string;
   width?: string;
-  flexBasis?: string;
+  flexBasis?: string | number;
+  flexGrow?: number;
   values?: Record<string, string>;
   alignCheckbox?: boolean;
   placeholder?: string;
@@ -490,6 +502,9 @@ export interface ScriptEventFieldSchema {
   fields?: ScriptEventFieldSchema[];
   inline?: boolean;
   allowedContexts?: ScriptEditorContextType[];
+  unitsField?: string;
+  unitsDefault?: UnitType;
+  unitsAllowed?: UnitType[];
   filter?: (value: unknown) => boolean;
   updateFn?: (
     newValue: unknown,